Triathlon Awarding
This program determines an award for triathlon athletes. It displays an understanding of: if/elif/else statements, casting, mathematical calculations, and logic reasoning. There are no error management implemented in this program, however this is an opportunity for improvement such as input format, and an error message to display wrong input i.e. non numerical value inputs.